Move interruptible executor to framework package
