+++ src/main/java/org/opentripplanner/standalone/server/MetricsLogging.java
@@ -19,8 +19,8 @@ import jakarta.inject.Inject;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ForkJoinPool;
-import org.opentripplanner.ext.parallelrouting.ParallelRouting;
 import org.opentripplanner.framework.application.OTPFeature;
+import org.opentripplanner.framework.concurrent.InterruptibleExecutor;
 import org.opentripplanner.graph_builder.issue.api.DataImportIssueSummary;
 import org.opentripplanner.raptor.configure.RaptorConfig;
 import org.opentripplanner.routing.algorithm.raptoradapter.transit.TripSchedule;
@@ -92,7 +92,7 @@ public class MetricsLogging {
 
     if (OTPFeature.ParallelRouting.isOn()) {
       new ExecutorServiceMetrics(
-        ParallelRouting.threadPool(),
+        InterruptibleExecutor.threadPool(),
         "parallelRouting",
         List.of(Tag.of("pool", "parallelRouting"))
       )
